+++ wok_code/wok_code/egg-info/CHANGELOG.rst
@@ -13,6 +13,7 @@
 * [FIX] odoo_translation.py: first character case
 * [FIX] odoo_translation.py: cache file format is Excel
 * [FIX] run_odoo_debug: path with heading space
+* [IMP] please test now can update account.account.xlsx
 
 2.0.12 (2023-08-29)
 ~~~~~~~~~~~~~~~~~~~